fix: abstraction of custom post type classes
