+++ includes/CPT/class-profile.php
@@ -12,7 +12,7 @@ use \Newspack\Govpack\Helpers;
 /**
  * Register and handle the "Profile" Custom Post Type
  */
-class Profile {
+class Profile extends \Newspack\Govpack\Post_Type {
 
 	/**
 	 * Valid profile formats.
@@ -42,14 +42,11 @@ class Profile {
 	 * WordPress Hooks
 	 */
 	public static function hooks() {
-		add_action( 'init', [ __CLASS__, 'register_post_type' ] );
+		parent::hooks();
 		add_action( 'cmb2_init', [ __CLASS__, 'add_profile_boxes' ] );
 		add_filter( 'wp_insert_post_data', [ __CLASS__, 'set_profile_title' ], 10, 3 );
 		add_action( 'edit_form_after_editor', [ __CLASS__, 'show_profile_title' ] );
 		add_filter( 'manage_edit-' . self::CPT_SLUG . '_sortable_columns', [ __CLASS__, 'sortable_columns' ] );
-		add_filter( 'manage_' . self::CPT_SLUG . '_posts_columns', [ __CLASS__, 'manage_columns' ] );
-		add_shortcode( self::SHORTCODE, [ __CLASS__, 'shortcode_handler' ] );
-		add_filter( 'body_class', [ __CLASS__, 'filter_body_class' ] );
 	}
 
 	/**
@@ -104,20 +101,6 @@ class Profile {
 	}
 
 	/**
-	 * Adds the post_type to array of supported post types.
-	 *
-	 * @param array $post_types   Array of post types.
-	 *
-	 * @return array
-	 */
-	public static function add_post_type( $post_types ) {
-		$post_types[] = static::CPT_SLUG;
-
-		return $post_types;
-	}
-
-
-	/**
 	 * Denote State, Party and Legislative Body columns as sortable.
 	 *
 	 * @param array $sortable_columns An array of sortable columns.
@@ -127,17 +110,6 @@ class Profile {
 		$sortable_columns[ 'taxonomy-' . \Newspack\Govpack\Tax\Party::TAX_SLUG ]           = 'Party';
 		$sortable_columns[ 'taxonomy-' . \Newspack\Govpack\Tax\LegislativeBody::TAX_SLUG ] = 'Legislative Body';
 		return $sortable_columns;
-	}
-
-	/**
-	 * Remove tags column from profile admin screen.
-	 *
-	 * @param string[] $columns The column header labels keyed by column ID.
-	 * @return array
-	 */
-	public static function manage_columns( $columns ) {
-		unset( $columns['tags'] );
-		return $columns;
 	}
 
 	/**
@@ -653,27 +625,6 @@ class Profile {
 		}
 
 		return $new_post;
-	}
-
-	/**
-	 * Add body classes depending on layout.
-	 *
-	 * @param array $classes CSS classes.
-	 *
-	 * @return array
-	 */
-	public static function filter_body_class( $classes ) {
-		if ( is_singular( self::CPT_SLUG ) ) {
-			$classes[] = 'archive';
-			$classes[] = 'feature-latest';
-
-			$key = array_search( 'single', $classes, true );
-			if ( false !== $key ) {
-				unset( $classes[ $key ] );
-			}
-		}
-
-		return $classes;
 	}
 
 	/**
